# [win-perf][proof-of-concept] run PS in a server locally