@chris-rock 2017-10-06
That is a great way, but I agree with Dom. I think we also could try to re-use the same powershell session and run multiple powershell commands via the same session. Maybe @mwrock has some ideas?