@sdelano 2017-10-06
@arlimus @chris-rock I 100% agree that opening an port for HTTP is not the way to go. I chose HTTP here because it was quick and easy in PowerShell and it proved that the method worked.

I've been working this week on taking the same approach and using IPC via pipes for the same effect: https://msdn.microsoft.com/en-us/library/windows/desktop/aa365574%28v=vs.85%29.aspx?f=255&MSPPError=-2147217396#base.using_pipes_for_ipc

This is very similar to what we're currently using for erlang -> ruby communication via the erlectricity (github.com/chef/erlectricity) project.