@jerryaldrichiii 2018-01-04
Closing this since work was implemented via https://github.com/chef/train/commit/d5dbab364a543070a04738a33e7866e116e2c99d